% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargarDatosDatalake.R
\name{cargarDatosDatalake}
\alias{cargarDatosDatalake}
\title{Cargar datos en el Datalake en AWS}
\usage{
cargarDatosDatalake(local_file, bucket_name, folder, s3_file_name)
}
\arguments{
\item{local_file}{Es el archivo local o dataframe que se almacenará en AWS, ej: 'estados_clientes.csv'}

\item{bucket_name}{Es el nombre del bucket en s3 de la zona del datalake donde se almacenarán los datos, ej: 'landing-zone-analitica'}

\item{folder}{Es el nombre de la carpeta dentro del bucket en s3, normalmente corresponde al nombre del proyecto, ej: 'movilidad_segura'}

\item{s3_file_name}{Es el nombre que tendra el archivo de datos dentro del datalake}
}
\value{
Un dataframe con los datos resultantes de la consulta, si entra un arreglo retorna un arreglo con los resultados.
}
\description{
Esta función permite cargar datos en el Datalake en AWS por medio de s3,
para cargar la información se requiere un access_key_id y un aws_secret_key_id
que es generado cuando se solicita el acceso al datalake de la compañia.
}
\examples{
cargarDatosDatalake('estados_clientes.csv','landing-zone-analitica','movilidad_segura','estados_clientes_ms.csv')

}
\keyword{cargar}
\keyword{datalake}
\keyword{datos}
