% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultaSqlDatalake.R
\name{consultaSqlDatalake}
\alias{consultaSqlDatalake}
\title{Consulta SQL a Athena}
\usage{
consultaSqlDatalake(query, path, soloEjecutar = FALSE)
}
\arguments{
\item{query}{Es el codigo de la consulta en SQl, ej: 'SELECT * FROM "database"."tablename"'}

\item{path}{Es la ruta del folder del proyecto dentro de un bucket en s3, ej 's3://landing-zone-analitica/movilidad_segura'}
}
\value{

}
\description{
Esta función realiza una consutla SQL al datalake por medio de AWS athena, devuelve en dataframe una tabla.
se requiere un access_key_id y un aws_secret_key_id que son generados
cuando se solicita el acceso al datalake, para ser establecidos a traves de aws CLI en el computador.
}
\examples{
consultaSqlDatalake('SELECT * FROM "databasetest"."tabletest"',"s3://landing-zone-analitica/movilidad_segura") )    


}
\keyword{athena}
\keyword{consulta}
\keyword{sql}
