% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargarDatosDatalake.R
\name{descargarDatosDatalake}
\alias{descargarDatosDatalake}
\title{Descargar datos del Datalake}
\usage{
descargarDatosDatalake(bucket, folder, s3_file_name, output_location)
}
\arguments{
\item{bucket}{Es el nombre del bucket en s3 de la zona del datalake donde se almacenar치n los datos, ej: 'landing-zone-analitica'}

\item{folder}{Es el nombre de la carpeta dentro del bucket en s3, normalmente corresponde al nombre del proyecto, ej: 'movilidad_segura'}

\item{s3_file_name}{Es el nombre que tendra el archivo de datos dentro del datalake, ej: 'estados_clientes_mv.csv'}

\item{output_location}{Es la localizaci칩n local donde se va a descargar el archivo o dataframe, o nombre con el que se guardar치, ej: 'estados_clientes.csv'}
}
\value{
Un dataframe con los datos resultantes de la consulta, si entra un arreglo retorna un arreglo con los resultados.
}
\description{
Esta funcion permite descargar un archivo de datos en el Datalake en AWS por medio de s3,
para descargar la informaci칩n se requiere un access_key_id y un aws_secret_key_id que son generados
cuando se solicita el acceso al datalake a AWS S3, para ser establecidos a traves de 
aws CLI en el computador.
}
\examples{
descargarDatosDatalake('landing-zone-analitica', 'movilidad_segura', 'estados_clientes_mv.csv', 'estados_clientes.csv')

}
\keyword{datos}
\keyword{descargar}
\keyword{s3}
