% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorarBucketDatalake.R
\name{explorarBucketDatalake}
\alias{explorarBucketDatalake}
\title{Explora dentro de un bucket del Datalake}
\usage{
explorarBucketDatalake(name_bucket)
}
\arguments{
\item{name_bucket}{nombre del bucket que se quiere consultar}
}
\value{
lista con los objetos existentes dentro del bucket consultado
}
\description{
Esta funcion permite explorar el contenido existente dentro de un bucket en el datalake
Solo devuelve el contenido si el usuario tiene permisos de lectura sobre el bucket consultado
}
\examples{
explorarBucketDatalake('landing-zone-analitica')

}
\keyword{bucket}
\keyword{datalake}
\keyword{explorar}
\keyword{s3}
