% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorarFolderBucketDatalake.R
\name{explorarFolderBucketDatalake}
\alias{explorarFolderBucketDatalake}
\title{Explora una carpeta dentro de un bucket del Datalake}
\usage{
explorarFolderBucketDatalake(name_bucket, name_folder)
}
\arguments{
\item{name_bucket}{nombre del bucket que se quiere consultar}

\item{name_folder}{nombre de la carpeta a consultar dentro del bucket}
}
\value{
lista con los objetos existentes dentro de la carpeta en el bucket consultado
}
\description{
Esta funcion permite explorar el contenido existente de una carpeta ubicada de un bucket en el datalake
Solo devuelve el contenido si el usuario tiene permisos de lectura sobre el bucket consultado
}
\examples{
explorarFolderBucketDatalake('landing-zone-analitica','Capacitacion')

}
\keyword{bucket}
\keyword{carpeta}
\keyword{datalake}
\keyword{explorar}
\keyword{folder}
\keyword{s3}
