% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listarBucketsDatalake.R
\name{listarBucketsDatalake}
\alias{listarBucketsDatalake}
\title{Listar nombres de buckets en el datalake}
\usage{
listarBucketsDatalake()
}
\value{
lista con todos los nombre de los buckets existentes en el datalake
}
\description{
Esta función lista los buckets del datalake existentes en AWS S3. Devuelve una lista con los nombres.
Se requiere un access_key_id y un aws_secret_key_id que son generados
cuando se solicita el acceso al datalake, para ser establecidos a traves de aws CLI en el computador.
}
\examples{
listarBucketsDatalake()
}
\keyword{buckets}
\keyword{listar}
\keyword{s3}
