% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryTeradataSeguro.R
\name{queryTeradataSeguro}
\alias{queryTeradataSeguro}
\title{Query a teradata de manera segura}
\usage{
queryTeradataSeguro(
  miSQL,
  usuario = NULL,
  contrasena = NULL,
  host = "10.205.24.12"
)
}
\arguments{
\item{miSQL}{Es el Query con el cual se hara la consulta a teradata}

\item{usuario}{es el nombre de usuario de la compañia asignado para acceder a teradata}

\item{contrasena}{es el password establecido para accedera a teradata}

\item{host}{direccion ip de host de teradata}
}
\value{
Un dataframe con los datos resultantes de la consulta, si entra un arreglo retorna un arreglo con los resultados
}
\description{
Esta función permite hacer consulta SQL a teradata mediante un query, usando el password ecriptado en 
dos archivos de llaves generados previamente mediante TJEncryptPassword.R
las llaves que contienen la clave encriptada deben estar en el mismo directorio que el notebook a ejecutar
}
\keyword{datos}
\keyword{extraer}
\keyword{teradata}
