% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformarCsvPqtDatalake.R
\name{transformarCsvPqtDatalake}
\alias{transformarCsvPqtDatalake}
\title{Transforma un archivo csv a parquet en el  datalake}
\usage{
transformarCsvPqtDatalake(
  archivo_fuente,
  archivo_destino,
  zona_dl_fuente = "",
  zona_dl_destino = "",
  ruta_proyecto_dl = "",
  rol_aws = "AWSGlueServiceRoleDefaultSura",
  aws_access_key_id = NULL,
  aws_secret_access_key = NULL
)
}
\arguments{
\item{archivo_fuente}{Es el dataframe que sufrira la transformación}

\item{archivo_destino}{es el nombre con el que quedara el archivo luego de transformarlo}

\item{zona_dl_fuente}{zona del datalake donde se encuentra el archivo que sera transformado}

\item{zona_dl_destino}{zona del datalake donde se almacenara el archivo transformado}

\item{ruta_proyecto_dl}{ruta de directorio dentro de un bucket del datalake}
}
\value{
La ruta en S3 del archivo resultante
}
\description{
Esta función permite transformar un archivo en formato .csv a formato .pqt usando 
la herramient glue del datalake. Ademas hace el movimiento entre las zonas del 
datalake si es necesario
}
\keyword{datalake}
\keyword{datos}
\keyword{extraer}
